extensive refactoring underway, bringing Merge project back into Crud project
