+++ Merge.SqlServer/Dialect.SqlServer.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{F8AC8DB0-DFE0-4FCE-A7B5-11C5316C7A62}</ProjectGuid>
+    <ProjectGuid>{6160BBDE-24A4-4964-961F-39A4A5D036D4}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Postulate.Orm.Merge.MySql</RootNamespace>
-    <AssemblyName>Postulate.Orm.Merge.MySql</AssemblyName>
+    <RootNamespace>Postulate.Orm.SqlServer</RootNamespace>
+    <AssemblyName>Postulate.Orm.SqlServer</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -31,6 +31,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="System" />
+    <Reference Include="System.ComponentModel.DataAnnotations" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
@@ -40,7 +41,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="MySql.cs" />
+    <Compile Include="SqlServer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -48,8 +49,15 @@
       <Project>{202c0f01-7ed0-48dc-a883-b27066d855fc}</Project>
       <Name>Merge</Name>
     </ProjectReference>
+    <ProjectReference Include="..\PostulateV1\Crud.csproj">
+      <Project>{74D32DED-6B61-4C06-99B7-4BF11A7325E0}</Project>
+      <Name>Crud</Name>
+    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
+    <PackageReference Include="AoReflectionHelper">
+      <Version>1.0.1</Version>
+    </PackageReference>
     <PackageReference Include="Dapper">
       <Version>1.50.2</Version>
     </PackageReference>
