--- Merge/Exceptions/ValidationException.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace Postulate.Orm.Merge.Exceptions
-{
-    public class ValidationException : Exception
-    {
-        public ValidationException(string message) : base(message)
-        {
-        }
-    }
-}
