+++ PostulateV1/Crud.csproj
@@ -52,6 +52,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Abstract\MergeAction.cs" />
     <Compile Include="Abstract\SeedData.cs" />
     <Compile Include="Abstract\SqlDb_ChangeTracking.cs" />
     <Compile Include="Abstract\SqlDb_Copy.cs" />
@@ -60,6 +61,7 @@
     <Compile Include="Abstract\SqlDb_Save.cs" />
     <Compile Include="Abstract\SqlDb_SaveMultiple.cs" />
     <Compile Include="Abstract\SqlDb_Update.cs" />
+    <Compile Include="Abstract\SqlScriptGenerator.cs" />
     <Compile Include="Attributes\ClusterAttribute.cs" />
     <Compile Include="Attributes\DefaultExpressionAttribute.cs" />
     <Compile Include="Attributes\DereferenceExpression.cs" />
@@ -68,9 +70,9 @@
     <Compile Include="Attributes\SchemaAttribute.cs" />
     <Compile Include="Attributes\TrackChangesAttribute.cs" />
     <Compile Include="Attributes\TrackDeletionsAttribute.cs" />
+    <Compile Include="Exceptions\ValidationException.cs" />
     <Compile Include="Extensions\ConnectionExtensions.cs" />
     <Compile Include="Interfaces\IDb.cs" />
-    <Compile Include="Interfaces\ISchemaMerge.cs" />
     <Compile Include="Interfaces\IUserProfile.cs" />
     <Compile Include="Abstract\SqlDb.cs" />
     <Compile Include="Abstract\Record.cs" />
@@ -90,46 +92,25 @@
     <Compile Include="Extensions\PropertyInfoSqlExtensions.cs" />
     <Compile Include="Extensions\TypeExtensions.cs" />
     <Compile Include="Exceptions\PermissionDeniedException.cs" />
-    <Compile Include="Merge\Action\AddColumn.cs" />
-    <Compile Include="Merge\Action\AlterColumn.cs" />
-    <Compile Include="Merge\Action\AlterForeignKey.cs" />
-    <Compile Include="Merge\Action\AlterPrimaryKey.cs" />
-    <Compile Include="Merge\Action\CreateForeignKey.cs" />
-    <Compile Include="Merge\Action\CreatePrimaryKey.cs" />
-    <Compile Include="Merge\Action\CreateSchema.cs" />
-    <Compile Include="Merge\Action\CreateTable.cs" />
-    <Compile Include="Merge\Action\DropColumn.cs" />
-    <Compile Include="Merge\Action\DropForeignKey.cs" />
-    <Compile Include="Merge\Action\DropPrimaryKey.cs" />
-    <Compile Include="Merge\Action\DropTable.cs" />
-    <Compile Include="Merge\Action\MergeAction.cs" />
-    <Compile Include="Merge\Action\MergeForeignKeyBase.cs" />
-    <Compile Include="Merge\Action\RenameColumn.cs" />
-    <Compile Include="Merge\Action\RenameTable.cs" />
-    <Compile Include="Merge\Action\SetPatchVersion.cs" />
-    <Compile Include="Merge\AddVersion.cs" />
-    <Compile Include="Merge\AlterColumnTypes.cs" />
-    <Compile Include="Merge\AlterForeignKeys.cs" />
-    <Compile Include="Merge\AlterPrimaryKeys.cs" />
-    <Compile Include="Merge\ColumnRef.cs" />
-    <Compile Include="Merge\CreateSchemas.cs" />
-    <Compile Include="Merge\DbObject.cs" />
-    <Compile Include="Merge\DropTables.cs" />
-    <Compile Include="Merge\ForeignKeyRef.cs" />
-    <Compile Include="Merge\Helpers.cs" />
-    <Compile Include="Merge\RenameColumns.cs" />
-    <Compile Include="Merge\RenameTables.cs" />
-    <Compile Include="Merge\SchemaMerge.cs" />
-    <Compile Include="Merge\SchemaMerge_CreateIfNotExists.cs" />
-    <Compile Include="Merge\SyncTablesAndColumns.cs" />
+    <Compile Include="Merge\Actions\AddColumn.cs" />
+    <Compile Include="Merge\Actions\AddForeignKey.cs" />
+    <Compile Include="Merge\Actions\AlterColumn.cs" />
+    <Compile Include="Merge\Actions\CreateTable.cs" />
+    <Compile Include="Merge\Actions\DropColumn.cs" />
+    <Compile Include="Merge\Actions\DropTable.cs" />
+    <Compile Include="Merge\Engine.cs" />
     <Compile Include="Models\ChangeHistory.cs" />
     <Compile Include="Models\ChangeHistoryRecord.cs" />
+    <Compile Include="Models\ColumnInfo.cs" />
+    <Compile Include="Models\ForeignKeyInfo.cs" />
+    <Compile Include="Models\LineRange.cs" />
+    <Compile Include="Models\MergeProgress.cs" />
+    <Compile Include="Models\TableInfo.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Models\PropertyChange.cs" />
     <Compile Include="Attributes\RenameFromAttribute.cs" />
     <Compile Include="Abstract\SqlDb_FindOrCreate.cs" />
     <Compile Include="SqlServerDb.cs" />
-    <Compile Include="Exceptions\ValidationException.cs" />
     <Compile Include="Extensions\StringExtensions.cs" />
     <Compile Include="Validation\RegexAttribute.cs" />
     <Compile Include="Validation\ValidationAttribute.cs" />
