+++ PostulateV1/Merge/Actions/AlterColumn.cs
@@ -1,11 +1,10 @@
-﻿using Postulate.Orm.Merge.Abstract;
-using Postulate.Orm.Merge.Extensions;
+﻿using Postulate.Orm.Abstract;
 using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Reflection;
 
-namespace Postulate.Orm.Merge.MergeActions
+namespace Postulate.Orm.Merge.Action
 {
     public class AlterColumn : MergeAction
     {
